requirements.txt
nltk
chemdataextractor2
matplotlib
pandas
rdkit
cuspy
pdfminer3k